middleware

+++ routes/web.php
@@ -21,22 +21,22 @@ Route::get('/login', [AccountController::class, 'viewLoginPage']);
 Route::post('/login', [AccountController::class, 'login']);
 Route::get('/register', [AccountController::class, 'viewRegisterPage']);
 Route::post('/register', [AccountController::class, 'register']);
-Route::get('/logout', [AccountController::class, 'viewLogoutPage']);
-Route::post('/logout', [AccountController::class, 'logout']);
+Route::get('/logout', [AccountController::class, 'viewLogoutPage'])->middleware('registered');
+Route::post('/logout', [AccountController::class, 'logout'])->middleware('registered');
 
-Route::get('/home', [ItemController::class, 'viewHomePage']);
-Route::get('/item/{item_id}', [ItemController::class, 'viewItemDetail']);
-Route::post('/buy-item/{item_id}', [ItemController::class, 'buyItem']);
+Route::get('/home', [ItemController::class, 'viewHomePage'])->middleware('registered');
+Route::get('/item/{item_id}', [ItemController::class, 'viewItemDetail'])->middleware('registered');
+Route::post('/buy-item/{item_id}', [ItemController::class, 'buyItem'])->middleware('registered');
 
-Route::get('/profile', [AccountController::class, 'viewProfilePage']);
-Route::patch('/update-profile', [AccountController::class, 'updateProfile']);
-Route::get('/update-success', [AccountController::class, 'viewSuccessPage']);
+Route::get('/profile', [AccountController::class, 'viewProfilePage'])->middleware('registered');
+Route::patch('/update-profile', [AccountController::class, 'updateProfile'])->middleware('registered');
+Route::get('/update-success', [AccountController::class, 'viewSuccessPage'])->middleware('registered');
 
-Route::get('/maintenance', [AccountController::class, 'viewMaintenancePage']);
-Route::get('/update-role/{account_id}', [AccountController::class, 'viewUpdateRolePage']);
-Route::patch('/update-role/{account_id}', [AccountController::class, 'updateRole']);
-Route::delete('/delete-account/{account_id}', [AccountController::class, 'deleteAccount']);
+Route::get('/maintenance', [AccountController::class, 'viewMaintenancePage'])->middleware('admin');
+Route::get('/update-role/{account_id}', [AccountController::class, 'viewUpdateRolePage'])->middleware('admin');
+Route::patch('/update-role/{account_id}', [AccountController::class, 'updateRole'])->middleware('admin');
+Route::delete('/delete-account/{account_id}', [AccountController::class, 'deleteAccount'])->middleware('admin');
 
-Route::get('/cart', [OrderController::class, 'viewCartPage']);
-Route::delete('/delete-from-cart/{order_id}', [OrderController::class, 'deleteItemFromCart']);
-Route::delete('/check-out', [OrderController::class, 'checkOut']);
+Route::get('/cart', [OrderController::class, 'viewCartPage'])->middleware('registered');
+Route::delete('/delete-from-cart/{order_id}', [OrderController::class, 'deleteItemFromCart'])->middleware('registered');
+Route::delete('/check-out', [OrderController::class, 'checkOut'])->middleware('registered');
